pip_requirements.txt
regex==2022.3.15
pandas==1.3.2
Babel==2.11.0
beautifulsoup4==4.10.0
numba==0.55.1
numpy==1.21.5
PyYAML==6.0